moved everything to resources, include instance library in certain cookbooks
