+++ resources/service.rb
@@ -1,7 +1,7 @@
 resource_name :icinga2_service
 provides :icinga2_service
 # enclosing_provider Chef::Provider::Icinga2Instance
-allowed_actions [:create, :delete, :nothing]
+property :cookbook, String, default: 'icinga2'
 property :display_name, String
 property :import, String
 property :host_name, String
@@ -29,6 +29,23 @@ property :action_url, String
 property :icon_image, String
 property :icon_image_alt, String
 property :custom_vars, Hash
-property :icinga2_template, [true, false], default: false
+property :template, [true, false], default: false
 property :template_support, TrueClass, default: true
 property :resource_properties, Array, default: %w(import display_name host_name groups check_command max_check_attempts check_period notification_period check_interval retry_interval enable_notifications enable_active_checks enable_passive_checks enable_event_handler enable_flapping enable_perfdata event_command flapping_threshold volatile zone command_endpoint notes notes_url action_url icon_image icon_image_alt custom_vars template)
+
+action :create do
+  object_template
+end
+
+action :delete do
+  object_template
+end
+
+action_class do
+  include Icinga2::Cookbook::Instances
+  def object_template
+    object_resources = []
+    object_resources << @new_resource
+    process_icinga2_resources(new_resource.resource_name.to_s.gsub('icinga2_', ''), new_resource.resource_properties, new_resource.template_support, object_resources)
+  end
+end
