+++ resources/timeperiod.rb
@@ -1,6 +1,8 @@
+include Icinga2::Cookbook::Instances
+unified_mode true
+
 resource_name :icinga2_timeperiod
 provides :icinga2_timeperiod
-allowed_actions [:create, :delete, :nothing]
 
 property :cookbook, String, default: 'icinga2'
 property :display_name, String
@@ -8,6 +10,23 @@ property :import, String
 property :ranges, Hash
 property :zone, String
 
-property :icinga2_template, [true, false], default: false
+property :template, [true, false], default: false
 property :template_support, TrueClass, default: true
 property :resource_properties, Array, default: %w(display_name ranges zone import template)
+
+action :create do
+  object_template
+end
+
+action :delete do
+  object_template
+end
+
+action_class do
+  include Icinga2::Cookbook::Instances
+  def object_template
+    object_resources = []
+    object_resources << @new_resource
+    process_icinga2_resources(new_resource.resource_name.to_s.gsub('icinga2_', ''), new_resource.resource_properties, new_resource.template_support, object_resources)
+  end
+end
